Change the order of installation methods in the README
The vimball method is the easiest way to install the module for people
that are not use to install vim plugins; it should appear first, before
the manual way which is the more painful and the more error prone
method.

The pathogen way is probably the cleanest way but it requires more
effort for users that are not already using it.

The commit change the order of the installation methods according to the
remarks above. It put first the vimball method, then the pathogen one,
and it put the manual way at the end. About the manual way, the commit
also add the "autoload" directory that needs to be copied and that was
not mentioned before.